separate version, init, subcommands
