+++ pkg/userFlags/subcommands.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/xaaha/hulak/pkg/envparser"
 	"github.com/xaaha/hulak/pkg/migration"
 	"github.com/xaaha/hulak/pkg/utils"
 )
@@ -17,6 +16,7 @@ var embeddedFiles embed.FS
 
 // User subcommands
 const (
+	Version = "version"
 	Migrate = "migrate"
 	// future subcommands
 	Init = "init"
@@ -46,6 +46,9 @@ func init() {
 // HandleSubcommands loops through all the subcommands
 func HandleSubcommands() error {
 	switch os.Args[1] {
+	case Version:
+		getVersion()
+		os.Exit(0)
 	case Migrate:
 		err := migrate.Parse(os.Args[2:])
 		if err != nil {
@@ -72,49 +75,6 @@ func HandleSubcommands() error {
 		utils.PrintRed("Enter a valid subcommand")
 		printHelpSubCommands()
 		os.Exit(1)
-	}
-	return nil
-}
-
-func handleInit() error {
-	apiOptionsFile := "apiOptions.yaml"
-	err := initialize.Parse(os.Args[2:])
-	if err != nil {
-		return fmt.Errorf("\n invalid subcommand %v", err)
-	}
-	// Check if -env flag is present
-	if *createEnvs {
-		envs := initialize.Args()
-		if len(envs) > 0 {
-			for _, env := range envs {
-				if err := envparser.CreateDefaultEnvs(&env); err != nil {
-					utils.PrintRed(err.Error())
-				}
-			}
-		} else {
-			utils.PrintWarning("No environment names provided after -env flag")
-		}
-	} else {
-		if err := envparser.CreateDefaultEnvs(nil); err != nil {
-			utils.PrintRed(err.Error())
-		}
-
-		content, err := embeddedFiles.ReadFile(apiOptionsFile)
-		if err != nil {
-			return err
-		}
-
-		root, err := utils.CreatePath(apiOptionsFile)
-		if err != nil {
-			return nil
-		}
-
-		if err := os.WriteFile(root, content, utils.FilePer); err != nil {
-			return fmt.Errorf("error on writing '%s' file: %s", apiOptionsFile, err)
-		}
-
-		utils.PrintGreen(fmt.Sprintf("Created '%s': %s", apiOptionsFile, utils.CheckMark))
-		utils.PrintGreen("Done " + utils.CheckMark)
 	}
 	return nil
 }
